move utils to tests
