+++ tests/contrib/aiobotocore/py35/test.py
@@ -5,7 +5,7 @@ from ddtrace.contrib.aiobotocore.patch import patch, unpatch
 from ..utils import aiobotocore_client
 from ...asyncio.utils import AsyncioTestCase, mark_asyncio
 from tests.tracer.test_tracer import get_dummy_tracer
-from ....utils import assert_span_http_status_code, assert_is_measured
+from .... import assert_is_measured, assert_span_http_status_code
 
 
 class AIOBotocoreTest(AsyncioTestCase):
