+++ tests/contrib/aiohttp/test_request.py
@@ -10,7 +10,7 @@ from ddtrace.contrib.aiohttp.patch import patch, unpatch
 from ddtrace.contrib.aiohttp.middlewares import trace_app
 
 from .utils import TraceTestCase
-from ...utils import assert_is_measured
+from ... import assert_is_measured
 
 
 class TestRequestTracing(TraceTestCase):
