+++ tests/contrib/botocore/test.py
@@ -10,8 +10,7 @@ from ddtrace.compat import stringify
 
 # testing
 from tests.opentracer.utils import init_tracer
-from ... import TracerTestCase
-from ...utils import assert_span_http_status_code, assert_is_measured
+from ... import TracerTestCase, assert_is_measured, assert_span_http_status_code
 
 
 class BotocoreTest(TracerTestCase):
