+++ tests/contrib/bottle/test_autopatch.py
@@ -4,7 +4,7 @@ import webtest
 
 from unittest import TestCase
 from tests.tracer.test_tracer import get_dummy_tracer
-from ...utils import assert_span_http_status_code
+from ... import assert_span_http_status_code
 
 from ddtrace import compat
