+++ tests/contrib/consul/test.py
@@ -6,8 +6,7 @@ from ddtrace.vendor.wrapt import BoundFunctionWrapper
 from ddtrace.contrib.consul.patch import patch, unpatch
 
 from ..config import CONSUL_CONFIG
-from ... import TracerTestCase
-from ...utils import assert_is_measured
+from ... import TracerTestCase, assert_is_measured
 
 
 class TestConsulPatch(TracerTestCase):
