+++ tests/contrib/dbapi/test_unit.py
@@ -4,8 +4,7 @@ from ddtrace import Pin
 from ddtrace.constants import ANALYTICS_SAMPLE_RATE_KEY
 from ddtrace.contrib.dbapi import FetchTracedCursor, TracedCursor, TracedConnection
 from ddtrace.span import Span
-from ... import TracerTestCase
-from ...utils import assert_is_measured, assert_is_not_measured
+from ... import TracerTestCase, assert_is_measured, assert_is_not_measured
 
 
 class TestTracedCursor(TracerTestCase):
