+++ tests/contrib/dogpile_cache/test_tracing.py
@@ -5,7 +5,7 @@ from ddtrace import Pin
 from ddtrace.contrib.dogpile_cache.patch import patch, unpatch
 
 from tests.tracer.test_tracer import get_dummy_tracer
-from ...utils import assert_is_measured
+from ... import assert_is_measured
 
 
 @pytest.fixture
