+++ tests/contrib/flask/test_errorhandler.py
@@ -1,7 +1,7 @@
 import flask
 
 from . import BaseFlaskTestCase
-from ...utils import assert_span_http_status_code
+from ... import assert_span_http_status_code
 
 
 class FlaskErrorhandlerTestCase(BaseFlaskTestCase):
