+++ tests/contrib/flask/test_hooks.py
@@ -2,7 +2,7 @@ from flask import Blueprint
 
 from ddtrace.ext import http
 from . import BaseFlaskTestCase
-from ...utils import assert_span_http_status_code
+from ... import assert_span_http_status_code
 
 
 class FlaskHookTestCase(BaseFlaskTestCase):
