+++ tests/contrib/flask/test_views.py
@@ -4,8 +4,7 @@ from ddtrace.compat import PY2
 from ddtrace.ext import http
 
 from . import BaseFlaskTestCase
-from ...utils import assert_span_http_status_code
-
+from ... import assert_span_http_status_code
 
 base_exception_name = 'builtins.Exception'
 if PY2:
