+++ tests/contrib/flask_cache/test.py
@@ -12,9 +12,8 @@ from flask import Flask
 # testing
 from tests.opentracer.utils import init_tracer
 from ..config import REDIS_CONFIG, MEMCACHED_CONFIG
-from ... import TracerTestCase
+from ... import TracerTestCase, assert_is_measured
 from ...util import assert_dict_issuperset
-from ...utils import assert_is_measured
 
 
 class FlaskCacheTest(TracerTestCase):
