+++ tests/contrib/mongoengine/test.py
@@ -14,10 +14,8 @@ from ddtrace.ext import mongo as mongox
 # testing
 from tests.opentracer.utils import init_tracer
 from ..config import MONGO_CONFIG
-from ...base import override_config
-from ... import TracerTestCase
+from ... import TracerTestCase, assert_is_measured
 from tests.tracer.test_tracer import get_dummy_tracer
-from ...utils import assert_is_measured
 
 
 class Artist(mongoengine.Document):
@@ -174,7 +172,7 @@ class MongoEngineCore(object):
         assert spans[0].get_metric(ANALYTICS_SAMPLE_RATE_KEY) is None
 
     def test_analytics_with_rate(self):
-        with override_config(
+        with TracerTestCase.override_config(
             'pymongo',
             dict(analytics_enabled=True, analytics_sample_rate=0.5)
         ):
@@ -186,7 +184,7 @@ class MongoEngineCore(object):
             assert spans[0].get_metric(ANALYTICS_SAMPLE_RATE_KEY) == 0.5
 
     def test_analytics_without_rate(self):
-        with override_config(
+        with TracerTestCase.override_config(
             'pymongo',
             dict(analytics_enabled=True)
         ):
