+++ tests/contrib/mysqldb/test_mysql.py
@@ -6,9 +6,8 @@ from ddtrace.contrib.mysqldb.patch import patch, unpatch
 
 from tests.opentracer.utils import init_tracer
 from ..config import MYSQL_CONFIG
-from ... import TracerTestCase
+from ... import TracerTestCase, assert_is_measured
 from ...util import assert_dict_issuperset
-from ...utils import assert_is_measured
 
 
 class MySQLCore(object):
