+++ tests/contrib/pylibmc/test.py
@@ -15,8 +15,7 @@ from ddtrace.ext import memcached
 # testing
 from ...opentracer.utils import init_tracer
 from ...contrib.config import MEMCACHED_CONFIG as cfg
-from ... import TracerTestCase
-from ...utils import assert_is_measured
+from ... import TracerTestCase, assert_is_measured
 
 
 class PylibmcCore(object):
