+++ tests/contrib/pymemcache/test_client_mixin.py
@@ -1,5 +1,4 @@
 # 3p
-import unittest
 import pymemcache
 
 # project
@@ -10,9 +9,7 @@ from ddtrace.ext import memcached as memcachedx, net
 from .utils import MockSocket
 
 from tests.tracer.test_tracer import get_dummy_tracer
-from ...base import override_config
-from ...utils import assert_is_measured
-
+from ... import TracerTestCase
 
 _Client = pymemcache.client.base.Client
 
@@ -20,7 +17,7 @@ TEST_HOST = 'localhost'
 TEST_PORT = 117711
 
 
-class PymemcacheClientTestCaseMixin(unittest.TestCase):
+class PymemcacheClientTestCaseMixin(TracerTestCase):
     """ Tests for a patched pymemcache.client.base.Client. """
 
     def get_spans(self):
@@ -35,7 +32,7 @@ class PymemcacheClientTestCaseMixin(unittest.TestCase):
         self.assertEqual(num_expected, len(spans))
 
         for span, resource, query in zip(spans, resources_expected, queries_expected):
-            assert_is_measured(span)
+            self.assert_is_measured(span)
             self.assertEqual(span.get_tag(net.TARGET_HOST), TEST_HOST)
             self.assertEqual(span.get_metric(net.TARGET_PORT), TEST_PORT)
             self.assertEqual(span.name, memcachedx.CMD)
@@ -151,7 +148,7 @@ class PymemcacheClientTestCaseMixin(unittest.TestCase):
         self.assertIsNone(spans[0].get_metric(ANALYTICS_SAMPLE_RATE_KEY))
 
     def test_analytics_with_rate(self):
-        with override_config(
+        with self.override_config(
             'pymemcache',
             dict(analytics_enabled=True, analytics_sample_rate=0.5)
         ):
@@ -164,7 +161,7 @@ class PymemcacheClientTestCaseMixin(unittest.TestCase):
         self.assertEqual(spans[0].get_metric(ANALYTICS_SAMPLE_RATE_KEY), 0.5)
 
     def test_analytics_without_rate(self):
-        with override_config(
+        with self.override_config(
             'pymemcache',
             dict(analytics_enabled=True)
         ):
