+++ tests/contrib/pymongo/test.py
@@ -14,10 +14,8 @@ from ddtrace.contrib.pymongo.patch import patch, unpatch
 # testing
 from tests.opentracer.utils import init_tracer
 from ..config import MONGO_CONFIG
-from ...base import override_config
-from ... import TracerTestCase
+from ... import TracerTestCase, assert_is_measured
 from tests.tracer.test_tracer import get_dummy_tracer
-from ...utils import assert_is_measured
 
 
 def test_normalize_filter():
@@ -317,7 +315,7 @@ class PymongoCore(object):
         assert spans[0].get_metric(ANALYTICS_SAMPLE_RATE_KEY) is None
 
     def test_analytics_with_rate(self):
-        with override_config(
+        with TracerTestCase.override_config(
             'pymongo',
             dict(analytics_enabled=True, analytics_sample_rate=0.5)
         ):
@@ -330,7 +328,7 @@ class PymongoCore(object):
             assert spans[0].get_metric(ANALYTICS_SAMPLE_RATE_KEY) == 0.5
 
     def test_analytics_without_rate(self):
-        with override_config(
+        with TracerTestCase.override_config(
             'pymongo',
             dict(analytics_enabled=True)
         ):
