+++ tests/contrib/pyramid/utils.py
@@ -13,8 +13,7 @@ from ddtrace.ext import http
 from .app import create_app
 
 from ...opentracer.utils import init_tracer
-from ... import TracerTestCase
-from ...utils import assert_span_http_status_code, assert_is_measured
+from ... import TracerTestCase, assert_is_measured, assert_span_http_status_code
 
 
 class PyramidBase(TracerTestCase):
