+++ tests/contrib/redis/test.py
@@ -9,8 +9,7 @@ from ddtrace.contrib.redis.patch import patch, unpatch
 from tests.opentracer.utils import init_tracer
 from ..config import REDIS_CONFIG
 from tests.tracer.test_tracer import get_dummy_tracer
-from ... import TracerTestCase
-from ...utils import assert_is_measured
+from ... import TracerTestCase, assert_is_measured
 
 
 def test_redis_legacy():
