+++ tests/contrib/rediscluster/test.py
@@ -5,9 +5,8 @@ from ddtrace import Pin
 from ddtrace.contrib.rediscluster.patch import patch, unpatch
 from ddtrace.contrib.rediscluster.patch import REDISCLUSTER_VERSION
 from ..config import REDISCLUSTER_CONFIG
-from ... import TracerTestCase
+from ... import TracerTestCase, assert_is_measured
 from tests.tracer.test_tracer import get_dummy_tracer
-from ...utils import assert_is_measured
 
 
 class TestRedisPatch(TracerTestCase):
