+++ tests/contrib/sqlalchemy/test_mysql.py
@@ -3,8 +3,7 @@ import pytest
 
 from .mixins import SQLAlchemyTestMixin
 from ..config import MYSQL_CONFIG
-from ... import TracerTestCase
-from ...utils import assert_is_measured
+from ... import TracerTestCase, assert_is_measured
 
 
 class MysqlConnectorTestCase(SQLAlchemyTestMixin, TracerTestCase):
