+++ tests/contrib/sqlalchemy/test_postgres.py
@@ -6,8 +6,7 @@ import pytest
 
 from .mixins import SQLAlchemyTestMixin
 from ..config import POSTGRES_CONFIG
-from ... import TracerTestCase
-from ...utils import assert_is_measured
+from ... import TracerTestCase, assert_is_measured
 
 
 class PostgresTestCase(SQLAlchemyTestMixin, TracerTestCase):
