+++ tests/tracer/test_span.py
@@ -11,8 +11,7 @@ from ddtrace.constants import (
 )
 from ddtrace.span import Span
 from ddtrace.ext import SpanTypes, errors, priority
-from tests import TracerTestCase
-from tests.utils import assert_is_measured, assert_is_not_measured
+from tests import TracerTestCase, assert_is_measured, assert_is_not_measured
 
 
 class SpanTestCase(TracerTestCase):
